Move collector logic to separate file
